[webview_flutter] Temporarily add web engine initialization API
ewk_init() and ewk_shutdown() of ewk webview used in webview_flutter are
designed to be called only once in a process.
If ewk_init() is called after ewk_shutdown() is called, SIGTRAP is called internally.

ewk_init() initializes the efl module and web engine's arguments data.
The efl module is initialized by default in OS, and Arguments data is also initialized through SetArguments() API,
so calling ewk_init() is not necessary.

Therefore, add a temporary API to call ewk_init() only when necessary.
This may be removed depending on the update status of chromium-efl.